+++ src/.vuepress/viewer/index.js
@@ -6,16 +6,24 @@ import shortcuts from "./config/shortcuts.js";
 import simple from "./config/simple.js";
 import state from "./config/state.js";
 import windowUI from "./config/windowUI.js";
+import defaultConfig from "./config/default.js";
+import guiLayout from "./config/guiLayout.js";
+import contextPlugins from "./config/contextPlugins.js";
+import globalComponents from "./config/globalComponents.js";
 
 const configs = new Map([
   ["contextMenu", contextMenu],
   ["hubs", hubs],
+  ["guiLayout", guiLayout],
+  ["contextPlugins", contextPlugins],
+  ["globalComponents", globalComponents],
   ["planAnnotations", planAnnotations],
   ["pluginUI", pluginUI],
   ["shortcuts", shortcuts],
   ["simple", simple],
   ["state", state],
   ["windowUI", windowUI],
+  ["default", defaultConfig],
 ]);
 
 async function makeViewer(config, id) {